Socket Class Created
I added transcribe_class_2.py which allows for Socket Based Transcription.l We still need the setup.cfg file for credentials. Right now it will write the audio every 10 seconds and stream results back from watson (which minimal output to the console.) I still need to update it so that information is pushed to a server and optionally set it up for batch processing.